+++ packages/core/src/v2/writer.ts
@@ -1,0 +1,7 @@
+import { Container } from "./container";
+
+export class GLTFWriter {
+    public static write(container: Container) {
+        throw new Error('Not implemented.');
+    }
+}
